Last setup password can now be retrieved
Password set on the first running of the application can now be retrieved on the next iteration of startups.

* This is done using the INITIAL_RUN variable flag that is set to True on the first run and set to False when Canda exits.
* The False value is stored in QSettings once the app exits.
* The app checks itself if it already had an initial startup to set the master key.
* If the INITIAL_RUN is set to False, the app will skip setting up the master key again to give way for the previously setup credentials.

This setting up of master key is done backend (no user interactions) but this will be implemented on a GUI wherein the user can set their own passwords.

On to a weekend break right now. Too many house chores to complete this weekend :smile: Coding session will continue on Monday. #happyweekend