Seamless flow from Setup dialog to Canda dialog
Well the 'if' statements works. I was able to confirm if the master key does match in the Setup dialog, login using the previously encrypted master key and voila, show the main Canda dilaog password manager #yehey :confetti_ball:

Next step is to get the user's preferred master key in the Setup dialog and encrypt it.