Fix INITIAL_RUN issue
The app can now determine itself if it initially setup a master key. This is done using a fix ACCOUNT_NAME save as main QSettings. The user no longer need to setup an account name, the app will just refer to the fix value equal to ACCOUNT_NAME. Frankly, I don't want to implement it this way but I have a deadline! :smile: I need to release at least a working first version. I'll just amend this current process in future versions of this app.

Also added a 'Set Master Key' button that will just show the 'Setup' dialog to update the current master key. Looking good so far and it works, it just needs some refining in the flow of dialogs shown and closed. Will continue working on this part next week.